+++ addOns/spider/CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Changed
+- Spider will not notify file seeds as found URI (Issue 7737).
 
 ## [0.11.0] - 2024-05-07
 ### Changed
@@ -13,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support for menu weights (Issue 8369)
+
 ### Fixed
 - Fix grammatical error in help content.